% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_minmax.R
\name{get_max_stat}
\alias{get_max_stat}
\title{get_max_stat}
\usage{
get_max_stat(seq, statfunc, n)
}
\arguments{
\item{seq}{A vector of categorical variables}

\item{statfunc}{The used statistic function}

\item{n}{The number of randomized samples}
}
\value{
Integer
}
\description{
An internal function that returns the approximate maximum value of a given statistic 
for randomized samples of the input vector of categorical variables.
It uses the Metropolis–Hastings algorithm for the approximation.
}
\keyword{internal}
