% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_minmax.R
\name{get_max_stat_heur_U}
\alias{get_max_stat_heur_U}
\title{get_max_stat_heur_U}
\usage{
get_max_stat_heur_U(seq)
}
\arguments{
\item{seq}{A vector of categorical values}
}
\value{
Integer
}
\description{
An internal function that returns the approximate maximum value of the U statistic 
for randomized samples of an input vector of categorical variables
using a heuristic rule.
}
\keyword{internal}
