% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_minmax.R
\name{get_max_stat_MH}
\alias{get_max_stat_MH}
\title{get_max_stat_MH}
\usage{
get_max_stat_MH(seq, statname, n)
}
\arguments{
\item{seq}{A vector of categorical values}

\item{statname}{The used statistic function}

\item{n}{The number of randomized samples}
}
\value{
Integer
}
\description{
An internal function that returns the approximate maximum value of a given statistic 
for randomized samples of an input vector of categorical variables
using the Metropolis–Hastings algorithm.
}
\keyword{internal}
