% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_minmax.R
\name{get_min_stat}
\alias{get_min_stat}
\title{get_min_stat}
\usage{
get_min_stat(seq, statfunc)
}
\arguments{
\item{seq}{A vector of categorical variables}

\item{statfunc}{The used statistic function}
}
\value{
Integer
}
\description{
An internal function that returns the exact minimum value of a given statistic 
for randomized samples of the input vector of categorical variables
}
\keyword{internal}
