% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_stats.R
\name{get_qualitative_spoiledness}
\alias{get_qualitative_spoiledness}
\title{get_qualitative_spoiledness}
\usage{
get_qualitative_spoiledness(seq)
}
\arguments{
\item{seq}{A vector of categorical variables}
}
\value{
The integer-valued statistic
}
\description{
An internal function that returns the qualitative spiledness statistic of
the input vector of categorical variables.
}
\keyword{internal}
