% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_probe.R
\name{get_statistics_for_randomized}
\alias{get_statistics_for_randomized}
\title{get_statistics_for_randomized}
\usage{
get_statistics_for_randomized(seq, statfunc, n)
}
\arguments{
\item{seq}{A vector of categorical variables}

\item{statfunc}{The used statistic function}

\item{n}{The number of randomized samples}
}
\value{
Vector of integers
}
\description{
An internal function that returns the list of statistics for 
n number of randomized versions of the input vector of categorical variables
}
\keyword{internal}
