% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_stats.R
\name{get_sum_of_rank_differences}
\alias{get_sum_of_rank_differences}
\title{get_sum_of_rank_differences}
\usage{
get_sum_of_rank_differences(seq)
}
\arguments{
\item{seq}{A vector of categorical variables}
}
\value{
The integer-valued statistic
}
\description{
An internal function that returns the sum of ranks statistic of
the input vector of categorical variables.
}
\keyword{internal}
