% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_stats.R
\name{get_T}
\alias{get_T}
\title{get_T}
\usage{
get_T(seq)
}
\arguments{
\item{seq}{A vector of categorical variables}
}
\value{
The integer-valued statistic
}
\description{
An internal function that returns the T statistic of
the input vector of categorical variables.
}
\keyword{internal}
