% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord.R
\docType{class}
\name{nomord-class}
\alias{nomord-class}
\title{An S4 class to represent numord test result}
\description{
An S4 class to represent numord test result
}
\section{Slots}{

\describe{
\item{\code{input}}{The vectorized input sequence}

\item{\code{p}}{p-value of the statistic test}
}}

