% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_probe.R
\name{nomord_probe}
\alias{nomord_probe}
\title{nomord_probe}
\usage{
nomord_probe(seq, statname)
}
\arguments{
\item{seq}{A vector of categorical variables}

\item{statname}{The used statistic function ("U" or "T")}
}
\value{
A nomord class S4 object with the following slots: input, statname, statvalue, althyp, p
}
\description{
Returns the probability p
}
\examples{
nomord_probe(c("B","A","A","C","C","B","A","A"), 'U')

}
