% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_minmax.R
\name{get_max_stat}
\alias{get_max_stat}
\title{get_max_stat}
\usage{
get_max_stat(seq, statname, n = -1)
}
\arguments{
\item{seq}{A vector of categorical values}

\item{n}{The number of randomized samples}

\item{statfunc}{The used statistic function}
}
\value{
Integer
}
\description{
An internal function that returns the approximate maximum value of a given statistic 
for randomized samples of an input vector of categorical variables
using either a heuristical rule or the Metropolis–Hastings algorithm.
}
\keyword{internal}
