% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_minmax.R
\name{get_max_stat}
\alias{get_max_stat}
\title{get_max_stat}
\usage{
get_max_stat(seq, statname)
}
\arguments{
\item{seq}{A vector of categorical values}

\item{statfunc}{The used statistic formula}
}
\value{
Integer
}
\description{
An internal function that returns the approximate maximum value of a given statistic 
for randomized samples of an input vector of categorical variables
using heuristic rules
}
\keyword{internal}
