% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_minmax.R
\name{get_min_stat}
\alias{get_min_stat}
\title{get_min_stat}
\usage{
get_min_stat(seq, statname)
}
\arguments{
\item{seq}{A vector of categorical values}

\item{statname}{The name of the statistic}
}
\value{
Integer
}
\description{
An internal function that returns the exact minimum value of a given statistic 
for randomized samples of an input vector of categorical variables
}
\keyword{internal}
