% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord_probe.R
\name{show,nomord-method}
\alias{show,nomord-method}
\title{nomord class generics method}
\usage{
\S4method{show}{nomord}(object)
}
\arguments{
\item{object}{An object}
}
\description{
nomord class generics method
}
\keyword{internal}
