% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomord.R
\name{test_association}
\alias{test_association}
\title{test_association}
\usage{
test_association(seq, statname = "U")
}
\arguments{
\item{seq}{A vector of categorical variables}

\item{statname}{The used statistic function}
}
\value{
result A nomord class S4 object
}
\description{
Returns the probability p
}
\examples{
test_association(c(2,3,1,2,2,3,3,2), 'A')

}
